rename reverseproxy middleware to decodeRepoPaths
