+++ td.server/src/config/routes.config.js
@@ -3,10 +3,10 @@ import express from 'express';
 import auth from '../controllers/auth.js';
 import bearer from './bearer.config.js';
 import configController from "../controllers/configcontroller";
+import decodeRepoPaths from './decodeRepoPaths.config.js';
 import googleProviderThreatmodelController from '../controllers/googleProviderThreatmodelController.js';
 import healthcheck from '../controllers/healthz.js';
 import homeController from '../controllers/homecontroller.js';
-import reverseproxy from './reverseproxy.config.js';
 import templateController from '../controllers/templateController.js';
 import threatmodelController from '../controllers/threatmodelcontroller.js';
 
@@ -50,22 +50,22 @@ const routes = (router) => {
 
     router.get('/api/threatmodel/repos', threatmodelController.repos);
     router.get('/api/threatmodel/:organisation/:repo/branches', threatmodelController.branches);
-    router.get('/api/threatmodel/:organisation/*repo/branches', reverseproxy.middleware, threatmodelController.branches);
+    router.get('/api/threatmodel/:organisation/*repo/branches', decodeRepoPaths.middleware, threatmodelController.branches);
     router.get('/api/threatmodel/:organisation/:repo/:branch/models', threatmodelController.models);
-    router.get('/api/threatmodel/:organisation/*repo/:branch/models', reverseproxy.middleware, threatmodelController.models);
+    router.get('/api/threatmodel/:organisation/*repo/:branch/models', decodeRepoPaths.middleware, threatmodelController.models);
     router.get('/api/threatmodel/:organisation/:repo/:branch/:model/data', threatmodelController.model);
-    router.get('/api/threatmodel/:organisation/*repo/:branch/:model/data', reverseproxy.middleware, threatmodelController.model);
+    router.get('/api/threatmodel/:organisation/*repo/:branch/:model/data', decodeRepoPaths.middleware, threatmodelController.model);
 
     router.post('/api/threatmodel/:organisation/:repo/:branch/createBranch', threatmodelController.createBranch);
-    router.post('/api/threatmodel/:organisation/*repo/:branch/createBranch', reverseproxy.middleware, threatmodelController.createBranch);
+    router.post('/api/threatmodel/:organisation/*repo/:branch/createBranch', decodeRepoPaths.middleware, threatmodelController.createBranch);
 
     // removed because of security denial of service concerns (denial of models)
     //router.delete('/api/threatmodel/:organisation/:repo/:branch/:model', threatmodelController.deleteModel);
 
     router.post('/api/threatmodel/:organisation/:repo/:branch/:model/create', threatmodelController.create);
-    router.post('/api/threatmodel/:organisation/*repo/:branch/:model/create', reverseproxy.middleware, threatmodelController.create);
+    router.post('/api/threatmodel/:organisation/*repo/:branch/:model/create', decodeRepoPaths.middleware, threatmodelController.create);
     router.put('/api/threatmodel/:organisation/:repo/:branch/:model/update', threatmodelController.update);
-    router.put('/api/threatmodel/:organisation/*repo/:branch/:model/update', reverseproxy.middleware, threatmodelController.update);
+    router.put('/api/threatmodel/:organisation/*repo/:branch/:model/update', decodeRepoPaths.middleware, threatmodelController.update);
 
     // Google Drive routes
     router.get('/api/googleproviderthreatmodel/folders', googleProviderThreatmodelController.folders);
