+++ td.server/test/config/routes.config.spec.js
@@ -1,17 +1,17 @@
-import express from 'express';
 import { expect } from 'chai';
+import express from 'express';
 import sinon from 'sinon';
 
 import auth from '../../src/controllers/auth.js';
 import bearer from '../../src/config/bearer.config.js';
+import configcontroller from "../../src/controllers/configcontroller";
+import decodeRepoPaths from '../../src/config/decodeRepoPaths.config.js';
+import { getMockApp } from '../mocks/express.mocks.js';
+import googleProviderThreatmodelController from '../../src/controllers/googleProviderThreatmodelController.js';
 import healthcheck from '../../src/controllers/healthz.js';
 import homeController from '../../src/controllers/homecontroller.js';
-import { getMockApp } from '../mocks/express.mocks.js';
 import routeConfig from '../../src/config/routes.config.js';
 import threatmodelController from '../../src/controllers/threatmodelcontroller.js';
-import configcontroller from "../../src/controllers/configcontroller";
-import googleProviderThreatmodelController from '../../src/controllers/googleProviderThreatmodelController.js';
-import reverseproxy from '../../src/config/reverseproxy.config.js';
 
 describe('config/routes.config.js routes', () => {
     let mockApp;
@@ -105,7 +105,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded GET /api/threatmodel/:organisation/*repo/branches', () => {
             expect(mockRouter.get).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/branches',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.branches
             );
         });
@@ -127,7 +127,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded GET /api/threatmodel/:organisation/*repo/:branch/models', () => {
             expect(mockRouter.get).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/:branch/models',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.models
             );
         });
@@ -142,7 +142,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded GET /api/threatmodel/:organisation/*repo/:branch/:model/data', () => {
             expect(mockRouter.get).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/:branch/:model/data',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.model
             );
         });
@@ -158,7 +158,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded POST /api/threatmodel/:organisation/*repo/:branch/createBranch', () => {
             expect(mockRouter.post).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/:branch/createBranch',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.createBranch
             );
         });
@@ -173,7 +173,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded POST /api/threatmodel/:organisation/*repo/:branch/:model/create', () => {
             expect(mockRouter.post).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/:branch/:model/create',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.create
             );
         });
@@ -188,7 +188,7 @@ describe('config/routes.config.js routes', () => {
         it('routes decoded PUT /api/threatmodel/:organisation/*repo/:branch/:model/update', () => {
             expect(mockRouter.put).to.have.been.calledWith(
                 '/api/threatmodel/:organisation/*repo/:branch/:model/update',
-                reverseproxy.middleware,
+                decodeRepoPaths.middleware,
                 threatmodelController.update
             );
         });
